+++ LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using LeaveManagement.Application.Exceptions;
 using LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using LeaveManagement.Application.Persistence.Contracts;
+using LeaveManagement.Domain;
 using MediatR;
 
 namespace LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands;
@@ -17,6 +19,10 @@ internal class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReq
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+        if (leaveRequest is null)
+        {
+            throw new NotFoundException(nameof(LeaveRequest), request.Id);
+        }
         await _leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
