+++ LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using LeaveManagement.Application.Exceptions;
 using LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
@@ -17,6 +18,11 @@ internal class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeCo
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.Get(request.Id);
+        if (leaveType is null)
+        {
+            throw new NotFoundException(nameof(LeaveTypes), request.Id);
+        }
+
         await _leaveTypeRepository.Delete(leaveType);
         return Unit.Value;
     }
